requirements.txt
peewee==2.8.0
pysqlcipher3==1.0.1
pycrypto==2.6.1